Rename: UserProfile Components의Profile 폴더로 이동
